votifier_service/pyproject.toml
[project]
name = "votifier-service"
version = "1.0.0"
description = "Votifier protocol listener for Minecraft voting sites"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=42.0.0",
    "mcrcon>=0.7.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
